requirements.txt
asgiref==3.5.2
Django==4.1
django-filter==22.1
django-rest-passwordreset==1.3.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
